[C] WIP: move internal NLS solve for GBODE into seperate file 	- create file gbode_internal_nls.c / .h for the internal solve strategy 		-> can now use proper BLAS routines as symbols do not conflict 		   with SUNDIALS BLAS (long int versions) 	- hide most of the implementation - public functions are: 		- gbInternalNlsAllocate, gbInternalNlsFree 		- gbInternalSolveNLS 		- gbInternalNlsGetScaledTolerances 		- gbInternalContraction (experimental contractive 		  error for FIRK with at least 1 real eigenvalue)
